refactor: split version sync logic and clarify TypeScript docs
