+++ scripts/sync-version-info/release.js
@@ -1,15 +1,13 @@
-#!/usr/bin/env node
-
 const fs = require('node:fs');
 const path = require('node:path');
 const { execFileSync } = require('node:child_process');
+const { normalizeVersion, compareVersions } = require('./version');
 
-const REPO_ROOT = path.resolve(__dirname, '..');
+const REPO_ROOT = path.resolve(__dirname, '../..');
 const RELEASES_PATH = path.join(REPO_ROOT, 'website/releases.json');
 const CORE_RULES_DIR = path.join(REPO_ROOT, 'internal/rules');
 const PLUGINS_DIR = path.join(REPO_ROOT, 'internal/plugins');
 const STABLE_VERSION_RE = /^\d+\.\d+\.\d+$/;
-const TYPESCRIPT_REPOSITORY = 'https://github.com/microsoft/TypeScript.git';
 const PLUGIN_GROUP_FALLBACKS = new Map([
   ['import', 'eslint-plugin-import'],
   ['jest', 'eslint-plugin-jest'],
@@ -21,19 +19,6 @@ const PLUGIN_GROUP_FALLBACKS = new Map([
   ['unicorn', 'eslint-plugin-unicorn'],
 ]);
 const pluginGroupByBlob = new Map();
-
-function normalizeVersion(version) {
-  return String(version).replace(/^v/, '');
-}
-
-function compareVersions(a, b) {
-  const aParts = normalizeVersion(a).split('.').map(Number);
-  const bParts = normalizeVersion(b).split('.').map(Number);
-  for (let i = 0; i < 3; i++) {
-    if (aParts[i] !== bParts[i]) return aParts[i] - bParts[i];
-  }
-  return 0;
-}
 
 function canonicalRuleId(group, rule) {
   return `${group}:${rule.replace(/_/g, '-')}`;
@@ -194,56 +179,6 @@ function writeReleases(data) {
   fs.writeFileSync(RELEASES_PATH, `${JSON.stringify(data, null, 2)}\n`);
 }
 
-function getTypeScriptBinding() {
-  // Read the pinned revision, including a staged compiler update. This also
-  // works when the submodule has not been initialized in this checkout.
-  const entry = getGitOutput(['ls-files', '--stage', '--', 'typescript-go']);
-  const commit = /^160000 ([0-9a-f]{40}) 0\ttypescript-go$/.exec(entry)?.[1];
-  if (!commit) throw new Error('No TypeScript submodule commit is recorded');
-  try {
-    getGitOutput([
-      'diff',
-      '--exit-code',
-      '--ignore-submodules=dirty',
-      '--',
-      'typescript-go',
-    ]);
-  } catch {
-    throw new Error('Stage the TypeScript submodule revision before syncing');
-  }
-
-  // Query upstream tags so a shallow submodule does not hide release tags.
-  // Annotated tags identify their commit through the peeled ref (^{}).
-  const tags = getGitOutput([
-    'ls-remote',
-    '--tags',
-    TYPESCRIPT_REPOSITORY,
-    'refs/tags/v*',
-  ]);
-  const refs = new Map();
-  for (const line of tags.split('\n')) {
-    const match =
-      /^([0-9a-f]{40})\s+refs\/tags\/v(\d+\.\d+\.\d+(?:-[\w.-]+)?)(\^\{\})?$/.exec(
-        line,
-      );
-    if (!match) continue;
-    const [, sha, version, peeled] = match;
-    const ref = refs.get(version) ?? {};
-    ref[peeled ? 'commit' : 'object'] = sha;
-    refs.set(version, ref);
-  }
-  const versions = [...refs]
-    .filter(([, ref]) => (ref.commit ?? ref.object) === commit)
-    .map(([version]) => version)
-    .sort(
-      (a, b) =>
-        compareVersions(a.split('-')[0], b.split('-')[0]) ||
-        Number(!a.includes('-')) - Number(!b.includes('-')) ||
-        a.localeCompare(b, 'en', { numeric: true }),
-    );
-  return { commit, releaseVersion: versions.at(-1) ?? null };
-}
-
 function syncFullHistory() {
   const stableTags = getStableTags();
   if (stableTags.length === 0) {
@@ -277,7 +212,7 @@ function syncFullHistory() {
   writeReleases(releases);
 }
 
-function syncCurrentVersion() {
+function syncCurrentVersion(getTypeScriptBinding) {
   const version = normalizeVersion(
     JSON.parse(fs.readFileSync(path.join(REPO_ROOT, 'package.json'), 'utf8'))
       .version,
@@ -333,7 +268,7 @@ function syncCurrentVersion() {
   const rules = currentRules.filter(
     (rule) => !previousRules.has(rule) && !assignedRules.has(rule),
   );
-  const typescript = getTypeScriptBinding();
+  const typescript = getTypeScriptBinding(getGitOutput);
   if (!published) {
     const release = { version, rules, typescript };
     if (targetIndex === -1) releases.push(release);
@@ -343,35 +278,4 @@ function syncCurrentVersion() {
   writeReleases(releases);
 }
 
-function printUsage() {
-  console.log('\nUsage:');
-  console.log(
-    '  pnpm sync:version-info       Sync main and the current stable release',
-  );
-  console.log(
-    '  pnpm sync:version-info full  Rebuild rule history, preserving TypeScript bindings',
-  );
-}
-
-function main() {
-  const args = process.argv.slice(2);
-  if (args.length > 1 || (args[0] && args[0] !== 'full')) {
-    throw new Error('The only supported argument is "full"');
-  }
-
-  if (args[0] === 'full') {
-    syncFullHistory();
-  } else {
-    syncCurrentVersion();
-  }
-  console.log(`Generated ${path.relative(REPO_ROOT, RELEASES_PATH)}.`);
-}
-
-try {
-  main();
-} catch (error) {
-  console.error(`Release sync failed: ${error.message}`);
-  process.exitCode = 1;
-} finally {
-  printUsage();
-}
+module.exports = { syncFullHistory, syncCurrentVersion };
