+++ website/docs/en/guide/_meta.json
@@ -64,6 +64,6 @@
   {
     "type": "file",
     "name": "typescript-versions",
-    "label": "TypeScript Compiler Versions"
+    "label": "TypeScript Versions Used by Rslint"
   }
 ]
